Refactor publisher locations
